Make user/group info grep more reliable
This commit uses -e when grepping for fields in the output of userinfo
or groupinfo. This should help avoid accidentally matching other fields
whose values contain the name of the field being grepped for (e.g. if
a user is named noname then grepping for name would match both the name
field and that user's home directory).